# Faster extrapolation