--- a/src/extrapolation/flat.jl
+++ b/src/extrapolation/flat.jl
@@ -1,16 +1,16 @@
 function extrap_prep{N,d}(::Type{Flat}, ::Val{N}, ::Val{d})
     xs_d = Symbol("xs_", d)
-    :($xs_d = clamp($xs_d, lbound(etp, $d), ubound(etp, $d)))
+    :($xs_d = clamp($xs_d, lbound(etp, $d, inds_etp[$d]), ubound(etp, $d, inds_etp[$d])))
 end
 
 function extrap_prep{N,d}(::Type{Flat}, ::Val{N}, ::Val{d}, ::Val{:lo})
     xs_d = Symbol("xs_", d)
-    :($xs_d = max($xs_d, lbound(etp, $d)))
+    :($xs_d = max($xs_d, lbound(etp, $d, inds_etp[$d])))
 end
 
 function extrap_prep{N,d}(::Type{Flat}, ::Val{N}, ::Val{d}, ::Val{:hi})
     xs_d = Symbol("xs_", d)
-    :($xs_d = min($xs_d, ubound(etp, $d)))
+    :($xs_d = min($xs_d, ubound(etp, $d, inds_etp[$d])))
 end
 
 function extrap_prep{N,d}(::Val{:gradient}, ::Type{Flat}, ::Val{N}, ::Val{d}, ::Val{:lo})
@@ -18,8 +18,8 @@ function extrap_prep{N,d}(::Val{:gradient}, ::Type{Flat}, ::Val{N}, ::Val{d}, ::
     xs_d = coords[d]
 
     quote
-        if $xs_d < lbound(etp, $d)
-            $xs_d = lbound(etp, $d)
+        if $xs_d < lbound(etp, $d, inds_etp[$d])
+            $xs_d = lbound(etp, $d, inds_etp[$d])
             gradient!(g, etp.itp, $(coords...))
             g[$d] = 0
             return g
@@ -32,8 +32,8 @@ function extrap_prep{N,d}(::Val{:gradient}, ::Type{Flat}, ::Val{N}, ::Val{d}, ::
     xs_d = coords[d]
 
     quote
-        if $xs_d > ubound(etp, $d)
-            $xs_d = ubound(etp, $d)
+        if $xs_d > ubound(etp, $d, inds_etp[$d])
+            $xs_d = ubound(etp, $d, inds_etp[$d])
             gradient!(g, etp.itp, $(coords...))
             g[$d] = 0
             return g