--- a/src/extrapolation/linear.jl
+++ b/src/extrapolation/linear.jl
@@ -2,8 +2,8 @@ function extrap_prep{N,d}(::Type{Linear}, ::Val{N}, ::Val{d}, ::Val{:lo})
     coords = [Symbol("xs_", k) for k in 1:N]
     xs_d = coords[d]
     quote
-        if $xs_d < lbound(etp.itp, $d)
-            $xs_d = lbound(etp.itp, $d)
+        if $xs_d < lbound(etp.itp, $d, inds_etp[$d])
+            $xs_d = lbound(etp.itp, $d, inds_etp[$d])
             return etp[$(coords...)] + gradient(etp, $(coords...))[$d] * (xs[$d] - $xs_d)
         end
     end
@@ -12,8 +12,8 @@ function extrap_prep{N,d}(::Type{Linear}, ::Val{N}, ::Val{d}, ::Val{:hi})
     coords = [Symbol("xs_", k) for k in 1:N]
     xs_d = coords[d]
     quote
-        if $xs_d > ubound(etp, $d)
-            $xs_d = ubound(etp, $d)
+        if $xs_d > ubound(etp, $d, inds_etp[$d])
+            $xs_d = ubound(etp, $d, inds_etp[$d])
             return etp[$(coords...)] + gradient(etp, $(coords...))[$d] * (xs[$d] - $xs_d)
         end
     end