--- a/src/extrapolation/reflect.jl
+++ b/src/extrapolation/reflect.jl
@@ -7,8 +7,8 @@ Next, if x is now in the upper part of this ''double-domain´´, reflect over th
 function extrap_prep_dim(::Type{Reflect}, d)
     xs_d = Symbol("xs_", d)
     quote
-        start = lbound(etp.itp, $d)
-        width = ubound(etp.itp, $d) - start
+        start = lbound(etp.itp, $d, inds_etp[$d])
+        width = ubound(etp.itp, $d, inds_etp[$d]) - start
 
         $xs_d = mod($xs_d - start, 2width) + start
         $xs_d > start + width && (xs_d = start + width - $xs_d)
@@ -18,9 +18,9 @@ end
 extrap_prep{N,d}(::Type{Reflect}, ::Val{N}, ::Val{d}) = extrap_prep_dim(Reflect, d)
 function extrap_prep{N,d}(::Type{Reflect}, ::Val{N}, ::Val{d}, ::Val{:lo})
     xs_d = Symbol("xs_", d)
-    :($xs_d < lbound(etp.itp, $d) && $(extrap_prep_dim(Reflect, d)))
+    :($xs_d < lbound(etp.itp, $d, inds_etp[$d]) && $(extrap_prep_dim(Reflect, d)))
 end
 function extrap_prep{N,d}(::Type{Reflect}, ::Val{N}, ::Val{d}, ::Val{:hi})
     xs_d = Symbol("xs_", d)
-    :($xs_d > ubound(etp.itp, $d) && $(extrap_prep_dim(Reflect, d)))
+    :($xs_d > ubound(etp.itp, $d, inds_etp[$d]) && $(extrap_prep_dim(Reflect, d)))
 end