@sglyon 2017-04-27
This looks _fantastic_. Great optimizing work.

I don't have time right now, but I will take a closer look at this before the end of the week (probably Friday).

If others look/review this and approve the PR, feel free to merge without me.